+++ arc/typing.py
@@ -17,6 +17,10 @@ class TypeProtocol(t.Protocol):
     def __convert__(cls, value, *args):
         ...
 
+    # @classmethod
+    # def __prompt__(cls, ctx, param):
+    #     ...
+
 
 CompareReturn = t.Literal[-1, 0, 1]
 
@@ -44,4 +48,9 @@ CallbackTime = t.Literal["before", "around", "after"]
 
 class CompletionProtocol(t.Protocol):
     def __completions__(self, info: CompletionInfo) -> list[Completion] | Completion:
+        ...
+
+
+class SupportsStr(t.Protocol):
+    def __str__(self) -> str:
         ...