+++ docs/examples/from_env.py
@@ -1,0 +1,12 @@
+import arc
+import os
+
+os.environ["VAL"] = "2"
+
+
+@arc.command()
+def command(val: int = arc.Argument(envvar="VAL")):
+    print(f"VAL: {val}")
+
+
+command()